Rename tests module
